refactor(analytics): migrate Notifications components to useAnalytics
