+++ app/components/Views/Notifications/Details/Fields/TransactionField.test.tsx
@@ -2,14 +2,14 @@
 import React from 'react';
 import { fireEvent, render } from '@testing-library/react-native';
 import TransactionField from './TransactionField';
-import * as useMetricsModule from '../../../../hooks/useMetrics';
+import * as useAnalyticsModule from '../../../../hooks/useAnalytics/useAnalytics';
 import * as useCopyClipboardModule from '../hooks/useCopyClipboard';
 import { ModalFieldType } from '../../../../../util/notifications';
 import MOCK_NOTIFICATIONS from '../../../../UI/Notification/__mocks__/mock_notifications';
 import { MetricsEventBuilder } from '../../../../../core/Analytics/MetricsEventBuilder';
 
 // Mock the required modules
-jest.mock('../../../../hooks/useMetrics');
+jest.mock('../../../../hooks/useAnalytics/useAnalytics');
 jest.mock('../hooks/useCopyClipboard');
 jest.mock('../../../../../../locales/i18n', () => ({
   strings: jest.fn((key) => key),
@@ -29,10 +29,10 @@ describe('TransactionField', () => {
   const mockCopyToClipboard = jest.fn();
 
   beforeEach(() => {
-    jest.spyOn(useMetricsModule, 'useMetrics').mockReturnValue({
+    jest.spyOn(useAnalyticsModule, 'useAnalytics').mockReturnValue({
       trackEvent: mockTrackEvent,
       createEventBuilder: MetricsEventBuilder.createEventBuilder,
-    } as unknown as useMetricsModule.IUseMetricsHook);
+    } as unknown as ReturnType<typeof useAnalyticsModule.useAnalytics>);
     jest
       .spyOn(useCopyClipboardModule, 'default')
       .mockReturnValue(mockCopyToClipboard);
