+++ app/components/Views/Notifications/Details/Footers/BlockExplorerFooter.test.tsx
@@ -6,16 +6,14 @@ import { fireEvent, render } from '@testing-library/react-native';
 import type { OnChainRawNotification } from '@metamask/notification-services-controller/notification-services';
 import { strings } from '../../../../../../locales/i18n';
 import BlockExplorerFooter from './BlockExplorerFooter';
-import {
-  MetaMetricsEvents,
-  useMetrics,
-} from '../../../../../components/hooks/useMetrics';
+import { useAnalytics } from '../../../../hooks/useAnalytics/useAnalytics';
+import { MetaMetricsEvents } from '../../../../../core/Analytics';
 import { ModalFooterType } from '../../../../../util/notifications/constants/config';
 import {
   MOCK_ON_CHAIN_NOTIFICATIONS,
   MOCK_FEATURE_ANNOUNCEMENT_NOTIFICATIONS,
 } from '../../../../UI/Notification/__mocks__/mock_notifications';
-import { MetricsEventBuilder } from '../../../../../core/Analytics/MetricsEventBuilder';
+import { AnalyticsEventBuilder } from '../../../../../util/analytics/AnalyticsEventBuilder';
 import { getNetworkDetailsFromNotifPayload } from '../../../../../util/notifications';
 
 jest.mock('react-native/Libraries/Linking/Linking', () => ({
@@ -29,13 +27,15 @@ jest.mock('../../../../../util/notifications', () => ({
   getNetworkDetailsFromNotifPayload: jest.fn(),
 }));
 
-jest.mock('../../../../../components/hooks/useMetrics');
+jest.mock('../../../../hooks/useAnalytics/useAnalytics', () => ({
+  useAnalytics: jest.fn(),
+}));
 
 const trackEventMock = jest.fn();
 
-(useMetrics as jest.MockedFn<typeof useMetrics>).mockReturnValue({
+(useAnalytics as jest.MockedFn<typeof useAnalytics>).mockReturnValue({
   trackEvent: trackEventMock,
-  createEventBuilder: MetricsEventBuilder.createEventBuilder,
+  createEventBuilder: AnalyticsEventBuilder.createEventBuilder,
   enable: jest.fn(),
   addTraitsToUser: jest.fn(),
   createDataDeletionTask: jest.fn(),
@@ -44,7 +44,7 @@ const trackEventMock = jest.fn();
   getDeleteRegulationId: jest.fn(),
   isDataRecorded: jest.fn(),
   isEnabled: jest.fn(),
-  getMetaMetricsId: jest.fn(),
+  getAnalyticsId: jest.fn(),
 });
 
 jest.mock('react-native/Libraries/Linking/Linking', () => ({
@@ -106,7 +106,7 @@ describe('BlockExplorerFooter', () => {
 
     expect(Linking.openURL).toHaveBeenCalled();
     expect(trackEventMock).toHaveBeenCalledWith(
-      MetricsEventBuilder.createEventBuilder(
+      AnalyticsEventBuilder.createEventBuilder(
         MetaMetricsEvents.NOTIFICATION_DETAIL_CLICKED,
       )
         .addProperties({
