+++ app/util/notifications/services/FCMService.test.ts
@@ -6,7 +6,8 @@ import { processNotification } from '@metamask/notification-services-controller/
 import { createMockNotificationEthSent } from '@metamask/notification-services-controller/notification-services/mocks';
 
 import FCMService from './FCMService';
-import { EVENT_NAME, MetaMetrics } from '../../../core/Analytics';
+import { EVENT_NAME } from '../../../core/Analytics';
+import { analytics } from '../../analytics/analytics';
 import { NativeModules, Platform } from 'react-native';
 
 // Firebase Mock
@@ -341,7 +342,7 @@ describe('FCMService - onClickPushNotificationWhenAppClosed', () => {
   };
 
   const arrangeMocks = () => {
-    const mockTrackEvent = jest.spyOn(MetaMetrics.getInstance(), 'trackEvent');
+    const mockTrackEvent = jest.spyOn(analytics, 'trackEvent');
     const firebaseMocks = arrangeFirebaseMocks();
     const nativeModuleMocks = arrangeNativeModuleMocks();
     return {
@@ -469,7 +470,7 @@ describe('FCMService - onClickPushNotificationWhenAppSuspended', () => {
   };
 
   const arrangeMocks = () => {
-    const mockTrackEvent = jest.spyOn(MetaMetrics.getInstance(), 'trackEvent');
+    const mockTrackEvent = jest.spyOn(analytics, 'trackEvent');
     const firebaseMocks = arrangeFirebaseMocks();
     const deeplinkCallback = jest.fn();
     return {
